+++ css/style.css
@@ -1437,3 +1437,6 @@ main {
   background-color: #ff0000;
 }
 
+.temp {
+  display: contents;
+}
