+++ index.html
@@ -133,7 +133,9 @@
       </section>
 
       <!-- Здесь будут изображения других пользователей -->
+      <section class="pictures  container  temp">
 
+      </section>
     </section>
 
     <!-- Полноэкранный показ изображения -->
