+++ js/utils.js
@@ -4,8 +4,10 @@ export const convertToNumber = (input) => parseInt(input.value, 10);
 
 export const debounce = (cb, timeoutDelay) => {
   let timeoutId;
+
   return (...rest) => {
     clearTimeout(timeoutId);
+
     timeoutId = setTimeout(() => cb.apply(this, rest), timeoutDelay);
   };
 };
